[MTGOSDK/Core/Reflection] Refactor namespace structure
